requirements.txt
aiofiles==24.1.0
alembic==1.13.3
annotated-types==0.7.0
coverage==7.6.1
greenlet==3.1.1
html5tagger==1.3.0
httptools==0.6.1
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
sanic==24.6.0
sanic-ext==23.12.0
sanic-routing==23.12.0
SQLAlchemy==2.0.35
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.20.0
websockets==13.1
